src/requirements.txt
uWSGI==2.0.15
amqp==2.1.4
appdirs==1.4.3
asn1crypto==0.22.0
Babel==2.4.0
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.6.0
billiard==3.5.0.2
boto3==1.4.4
botocore==1.5.35
bs4==0.0.1
celery==4.0.2
certifi==2017.1.23
cffi==1.10.0
cryptography==1.8.1
DAWG-Python==0.7.2
Django==1.11
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-cors-headers==1.2.1
django-mptt==0.8.7
django-storages==1.5.2
docopt==0.6.2
docutils==0.13.1
elasticsearch==5.3.0
elasticsearch-dsl==5.2.0
enum34==1.1.6
Fabric==1.13.1
flower==0.9.1
futures==3.0.5
geoip2==2.5.0
gevent==1.2.1
greenlet==0.4.12
gunicorn==19.7.1
idna==2.5
ipaddress==1.0.18
jmespath==0.9.2
jws==0.1.3
kombu==4.0.2
maxminddb==1.3.0
olefile==0.44
packaging==16.8
paramiko==2.1.2
phonenumbers==8.5.0
Pillow==4.0.0
psycopg2==2.7.1
pyasn1==0.2.3
pycparser==2.17
PyJWT==1.4.2
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pyparsing==2.2.0
python-dateutil==2.6.0
python-jwt==2.0.1
python-memcached==1.58
pytz==2017.2
redis==2.10.6
requests==2.11.1
s3transfer==0.1.10
six==1.10.0
tornado==4.2
ua-parser==0.7.3
urllib3==1.20
user-agents==1.1.0
vine==1.1.3
xlrd==1.0.0
xlwt==1.1.2
pytils==0.3
